Don't wrap included URLs with ``multiurl``. It can break middleware.
E.g. CSRF middleware looks for a ``csrf_exempt`` attribute on the view, which
might be set on an included view but is not set on the multiurl view that
contains it.

It's also unecessary. If none of the included URLs match, subsequent URLs in the
root URLconf will still be checked.

For some reason this doesn't seem to be the case for the admin app, so it still
needs to come after the admin docs app (as they share a common prefix).